+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Extending the adopted spec, each change should have a link to its corresponding 
 
 ## [Unreleased]
 
+### Added
+
+- Support for Forseti Real-Time Enforcer VM private IP address. [#180]
+
 ## [v2.2.0] - 2019-06-20
 
 ### Added
@@ -192,6 +196,7 @@ Extending the adopted spec, each change should have a link to its corresponding 
 [v2.1.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v2.0.0...v2.1.0
 [v2.2.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v2.1.0...v2.2.0
 
+[#180]: https://github.com/forseti-security/terraform-google-forseti/pull/180
 [#174]: https://github.com/forseti-security/terraform-google-forseti/pull/174
 [#170]: https://github.com/forseti-security/terraform-google-forseti/pull/170
 [#163]: https://github.com/forseti-security/terraform-google-forseti/pull/163